+++ src/inspect_ai/_view/www/src/app/log-list/LogsFilterInput.module.css
@@ -1,0 +1,5 @@
+.filterInput {
+  font-size: 0.8em;
+  margin: 0em;
+  color: var(--bs-secondary);
+}
