+++ src/inspect_ai/_view/www/src/app/log-list/LogsPanel.module.css
@@ -2,7 +2,7 @@
   height: 100vh;
   width: 100vw;
   display: grid;
-  grid-template-rows: max-content max-content max-content 1fr max-content;
+  grid-template-rows: max-content max-content 1fr max-content;
 }
 
 .list {
