+++ src/inspect_ai/_view/www/src/app/log-list/LogsPanel.tsx
@@ -11,8 +11,8 @@ import { logUrl, useLogRouteParams } from "../routing/url";
 import { LogListGrid } from "./grid/LogListGrid";
 import { FileLogItem, FolderLogItem } from "./LogItem";
 import { LogListFooter } from "./LogListFooter";
+import { LogsFilterInput } from "./LogsFilterInput";
 import styles from "./LogsPanel.module.css";
-import { LogsToolbar } from "./LogsToolbar";
 
 const rootName = (relativePath: string) => {
   const parts = relativePath.split("/");
@@ -116,8 +116,10 @@ export const LogsPanel: FC<LogsPanelProps> = () => {
 
   return (
     <div className={clsx(styles.panel)}>
-      <Navbar />
-      <LogsToolbar />
+      <Navbar>
+        <LogsFilterInput />
+      </Navbar>
+
       <ProgressBar animating={loading || headersLoading} />
       <div className={clsx(styles.list, "text-size-smaller")}>
         <LogListGrid items={logItems} />
