--- src/inspect_ai/_view/www/src/app/log-list/LogsToolbar.module.css
@@ -1,30 +1,0 @@
-.toolbar {
-  border-bottom: solid var(--bs-border-color) 1px;
-  background: var(--bs-light);
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-}
-
-.right {
-  display: grid;
-  grid-auto-columns: max-content;
-  grid-auto-flow: column;
-  justify-content: end;
-  align-items: center;
-  column-gap: 0.5em;
-}
-
-.left {
-  display: grid;
-  grid-auto-columns: max-content;
-  grid-auto-flow: column;
-  justify-content: start;
-  align-items: center;
-  column-gap: 0.5em;
-}
-
-.filterInput {
-  font-size: 0.8em;
-  margin: 0.3em;
-  color: var(--bs-secondary);
-}
